Move unref from ObjectBase class into a separate class
